Moves widget-detail layout to layout/app
