+++ layout/app/widget-detail/index.js
