+++ layout/app/widget-detail/widget-detail-header/component.js
