+++ layout/app/widget-detail/widget-detail-header/index.js
